+++ include/http-server/BuildingSystem/HttpAdapter.h
