+++ src/BuildingSystem/CompositePattern/Component.cpp
@@ -3,8 +3,8 @@
 //
 
 #include <iostream>
-#include "http-server/Building/CompositePattern/Room.h"
-#include "http-server/Building/CompositePattern/Composite.h"
+#include "http-server/BuildingSystem/CompositePattern/Room.h"
+#include "http-server/BuildingSystem/CompositePattern/Composite.h"
 
 
 namespace BuildingSystem
