+++ src/BuildingSystem/CompositePattern/Equipment.cpp
@@ -2,9 +2,9 @@
 // Created by Julia on 2019-12-22.
 //
 
-#include "http-server/Building/CompositePattern/Component.h"
-#include "http-server/Building/Equipment.h"
-#include "http-server/Building/Equipment.h"
+#include "http-server/BuildingSystem/CompositePattern/Component.h"
+#include "http-server/BuildingSystem/CompositePattern/Equipment.h"
+#include "http-server/BuildingSystem/CompositePattern/Equipment.h"
 
 namespace BuildingSystem
 {
